Add support for different locale storage and train
Prior to this change, the DB level supported only a single locale -
English and there was an issue with the tapping on the context during
training.

This change stores the entry locale in the DB and fixes tapping on the
context field. Also, tapping on the context and "speak" feature supports
different locales.